fix(tests): clippy -D warnings 修复与全量负载下瞬态 flake 加固
- session_index 组合调用去掉 needless borrow；TUI 渲染去掉冗余 cast
- 持久化失败测试恢复可写夹具标注 permissions_set_readonly_false 豁免
  （临时目录内恢复可写，Unix 放宽只影响该临时文件）
- steer_transport followUp 第三请求等待界 5s→20s：独立运行恒绿，
  全量并行时曾出现瞬态超时；放宽后仍快速失败

+++ crates/cli/src/tui/app.rs
@@ -515,7 +515,7 @@ impl TuiApp {
     fn flow_metrics(&self) -> (usize, usize) {
         let width = self.last_flow_width.unwrap_or(80);
         let total: usize = self.transcript.row_counts(width).iter().sum::<usize>()
-            + self.transcript.live_row_count(width as u16);
+            + self.transcript.live_row_count(width);
         (total, self.last_viewport_rows.max(1))
     }
